created Orders partially

+++ BiteMeProtoSrv/src/common/Order.java
@@ -5,20 +5,20 @@ public class Order {
 	private String order_address;
 	private int order_list_number;
 	private int order_number;
-	private int total_price;
+	private double total_price;
 	
 	public Order(String name_of_restaurant, String order_address, int order_list_number, int order_number,
-			int total_price) {
+			double total_price) {
 		this.name_of_restaurant = name_of_restaurant;
 		this.order_address = order_address;
 		this.order_list_number = order_list_number;
 		this.order_number = order_number;
 		this.total_price = total_price;
 	}
-	public int getTotal_price() {
+	public double getTotal_price() {
 		return total_price;
 	}
-	public void setTotal_price(int total_price) {
+	public void setTotal_price(double total_price) {
 		this.total_price = total_price;
 	}
 	public int getOrder_number() {
